requirements.txt
flask==2.3.2
flask-sqlalchemy==3.0.3
flask-migrate==4.0.4
flask-login==0.6.2
flask-wtf==1.1.1
flask-cors==4.0.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
dropbox==11.36.0
PyJWT==2.7.0
werkzeug==2.3.4
pillow==9.5.0
pyqt5==5.15.9
kivy==2.2.1